+++ eras/babbage/impl/CHANGELOG.md
@@ -3,8 +3,7 @@
 ## 1.15.0.0
 
 * Add `FromJSON` instance for `BabbageTxOut era`
-* Change `transTxRedeemers` to accept a `UTxO era` argument
-* Change `toPlutusV2Args` to accept `LedgerTxInfo era` and `ScriptHash` arguments instead of `ProtVer` and `Maybe (Data era)`
+* Change `toPlutusV2Args` to accept `LedgerTxInfo era` argument instead of `ProtVer` and `Maybe (Data era)`
 * Rename `transRedeemerPtr` to `transRedeemerPointerV2V3`
 * Add `EncCBOR`, `ToCBOR` for `Block`
 * Add `DecCBOR` instances for `Annotator Block`